# backend/pkg/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  models.Container:
    properties:
      created_at:
        type: string
      flow_id:
        minimum: 0
        type: integer
      id:
        minimum: 0
        type: integer
      image:
        type: string
      local_dir:
        type: string
      local_id:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    required:
    - flow_id
    - image
    - local_dir
    - local_id
    - name
    - status
    - type
    type: object
...
  services.containers:
    properties:
      containers:
        items:
          $ref: '#/definitions/models.Container'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    email: team@pennull.com
    name: penNULL Development Team
    url: https://pennull.com
  description: Swagger API for Penetration Testing Advanced General Intelligence penNULL.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: penNULL Swagger API
  version: "1.0"
paths:
...
  /containers/:
    get:
      parameters:
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"...","operator":"..."}
            field is the unique identifier of the table column, different for each endpoint
            value should be integer or string or array type, "value":123 or "value":"string" or "value":[123,456]
            operator value should be one of <,<=,>=,>,=,!=,like,not like,in
            default operator value is 'like' or '=' if field is 'id' or '*_id' or '*_at'
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 1000, -1 means unlimited)
        in: query
        maximum: 1000
        minimum: -1
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
            order is required if prop is not empty
        in: query
        items:
          type: string
        name: sort[]
        type: array
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: containers list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.containers'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting containers not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting containers
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve containers list
      tags:
      - Containers
...
  /flows/{flowID}/containers/:
    get:
      parameters:
      - description: flow id
        in: path
        minimum: 0
        name: flowID
        required: true
        type: integer
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"...","operator":"..."}
            field is the unique identifier of the table column, different for each endpoint
            value should be integer or string or array type, "value":123 or "value":"string" or "value":[123,456]
            operator value should be one of <,<=,>=,>,=,!=,like,not like,in
            default operator value is 'like' or '=' if field is 'id' or '*_id' or '*_at'
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 1000, -1 means unlimited)
        in: query
        maximum: 1000
        minimum: -1
        name: pageSize
        type: integer
      - collectionFormat: multi
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
            order is required if prop is not empty
        in: query
        items:
          type: string
        name: sort[]
        type: array
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: containers list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.containers'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting containers not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting containers
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve containers list by flow id
      tags:
      - Containers
  /flows/{flowID}/containers/{containerID}:
    get:
      parameters:
      - description: flow id
        in: path
        minimum: 0
        name: flowID
        required: true
        type: integer
      - description: container id
        in: path
        minimum: 0
        name: containerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: container info received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Container'
              type: object
        "403":
          description: getting container not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: container not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting container
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve container info by id and flow id
      tags:
      - Containers
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"